+++ cli/internal_testsuites/typescript/src/test_helpers/enclave_setup.ts
@@ -1,5 +1,5 @@
-import { EnclaveContext, EnclaveID } from "kurtosis-core-sdk"
-import { KurtosisContext } from "kurtosis-engine-api-lib"
+import { EnclaveContext, EnclaveID } from "kurtosis-sdk"
+import { KurtosisContext } from "kurtosis-sdk"
 import {Result, err, ok} from "neverthrow"
 import log from "loglevel";