+++ cli/internal_testsuites/typescript/src/test_helpers/test_helpers.ts
@@ -8,7 +8,7 @@ import {
     PortSpec,
     PortProtocol,
     FilesArtifactUUID,
-} from "kurtosis-core-sdk";
+} from "kurtosis-sdk";
 import * as datastoreApi from "example-datastore-server-api-lib";
 import * as serverApi from "example-api-server-api-lib";
 import { err, ok, Result } from "neverthrow";