+++ cli/internal_testsuites/typescript/src/testsuite/basic_datastore_test/basic_datastore.test.ts
@@ -1,5 +1,5 @@
 import { UpsertArgs, ExistsArgs, GetArgs, ExistsResponse, GetResponse } from "example-datastore-server-api-lib";
-import { ServiceID } from "kurtosis-core-sdk"
+import { ServiceID } from "kurtosis-sdk"
 import * as grpc from "@grpc/grpc-js"
 import * as google_protobuf_empty_pb from "google-protobuf/google/protobuf/empty_pb";
 import log from "loglevel"