+++ cli/internal_testsuites/typescript/src/testsuite/network_soft_partition_test/network_soft_partition.test.ts
@@ -8,8 +8,8 @@ import {
     ServiceID, 
     SoftPartitionConnection,
     UnblockedPartitionConnection
-} from "kurtosis-core-sdk";
-import { PartitionConnection } from "kurtosis-core-sdk/build/lib/enclaves/partition_connection";
+} from "kurtosis-sdk";
+import { PartitionConnection } from "kurtosis-sdk/build/core/lib/enclaves/partition_connection";
 import log from "loglevel";
 import { ok, Result, err } from "neverthrow";